backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic
python-dotenv
requests
httpx
sse-starlette
bcrypt
python-jose[cryptography]
passlib[bcrypt]
ollama
yfinance
edge-tts
ffmpeg-python
phonemizer>=3.2.1
